Add simple logging to the scrapper (debug to stdout, info and higher goes to the default scrapper.log)
Add option to specify destination directory for the json "database". The
scrapper.log will be located in the same directory.
Option is `-d` or `--db_dir`
By default the database will be saved in scrapper/db

Reformat date to a unique format

Handle restarting the scrapper. It will not query for expansions it
already has